Move main form to correct namespace
